backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
pydantic[email]
pydantic-settings
bcrypt
python-jose[cryptography]
python-multipart
aiofiles
python-dotenv
